requirements.txt
Flask>=0.10.1
Flask-Bootstrap>=3.1.1.2
Flask-HTTPAuth>=2.2.1
Flask-Login>=0.2.11
Flask-Mail>=0.9.0
Flask-Migrate>=1.2.0
Flask-Moment>=0.4.0
Flask-PageDown>=0.1.4
Flask-SQLAlchemy>=1.0
Flask-Script>=2.0.5
Flask-WTF>=0.10.0
ForgeryPy>=0.1
Jinja2>=2.7.3
Mako>=1.0.0
Markdown>=2.4.1
MarkupSafe>=0.18
Pygments>=1.6
SQLAlchemy>=0.9.7
WTForms>=2.0.1
Werkzeug>=0.9.6
alembic>=0.6.5
backports.ssl-match-hostname>=3.4.0.2
bleach>=1.4
blinker>=1.3
colorama>=0.3.1
coverage>=3.7.1
dill>=0.2.1
futures>=2.1.6
html5lib>=1.0b3
httpie>=0.8.0
ipython>=2.2.0
itsdangerous>=0.24
matplotlib>=1.4.0
numexpr>=2.4
numpy>=1.8.1
pandas>=0.14.1
pyparsing>=2.0.2
pyreadline>=2.0
python-dateutil>=1.5
pytz>=2014.4
pyzmq>=14.3.1
requests>=2.3.0
scikit-learn>=0.15.2
scipy>=0.14.0
six>=1.7.3
table>=0.0.4
tables>=3.1.1
tornado>=4.0.1
wheel>=0.24.0
wsgiref>=0.1.2
xlrd>=0.9.3
pysaml2>=4.0.3
pyparsing>=2.1.1